SDIT-2569 Turn on contact person profile details in preprod, and some refactoring
