+++ helm_deploy/values-preprod.yaml
@@ -30,8 +30,6 @@ generic-service:
     FEATURE_EVENT_OFFENDER_SENTENCES-INSERTED: false
     FEATURE_EVENT_OFFENDER_SENTENCES-DELETED: false
     FEATURE_EVENT_OFFENDER_SENTENCES-UPDATED: false
-    FEATURE_EVENT_PERSONALRELATIONSHIPS_OFFENDER_PHYSICAL_DETAILS-CHANGED: false
-    FEATURE_EVENT_PERSONALRELATIONSHIPS_PRISON-OFFENDER-EVENTS_PRISONER_BOOKING_MOVED: false
     FEATURE_EVENT_OFFENDER_VISIT_BALANCE_ADJS-INSERTED: false
     FEATURE_EVENT_OFFENDER_VISIT_BALANCE_ADJS-UPDATED: false
     FEATURE_EVENT_OFFENDER_VISIT_BALANCE_ADJS-DELETED: false
