+++ helm_deploy/values-prod.yaml
@@ -69,7 +69,7 @@ generic-service:
     FEATURE_EVENT_PERSONALRELATIONSHIPS_PERSON_IDENTIFIERS-UPDATED: false
     FEATURE_EVENT_PERSONALRELATIONSHIPS_PERSON_IDENTIFIERS-DELETED: false
     FEATURE_EVENT_PERSONALRELATIONSHIPS_OFFENDER_PHYSICAL_DETAILS-CHANGED: false
-    FEATURE_EVENT_PERSONALRELATIONSHIPS_PRISON-OFFENDER-EVENTS_PRISONER_BOOKING_MOVED: false
+    FEATURE_EVENT_PERSONALRELATIONSHIPS_PROFILEDETAILS_PRISON-OFFENDER-EVENTS_PRISONER_BOOKING_MOVED: false
     FEATURE_EVENT_PERSONCONTACTS_PRISON-OFFENDER-EVENTS_PRISONER_MERGED: false
     FEATURE_EVENT_PERSONCONTACTS_PRISON-OFFENDER-EVENTS_PRISONER_BOOKING_MOVED: false
     FEATURE_EVENT_PERSONCONTACTS_PRISONER-OFFENDER-SEARCH_PRISONER_RECEIVED: false
