+++ src/main/kotlin/uk/gov/justice/digital/hmpps/prisonerfromnomismigration/personalrelationships/profiledetails/ContactPersonProfileDetailsBookingMovedService.kt
@@ -7,9 +7,9 @@ import uk.gov.justice.digital.hmpps.prisonerfromnomismigration.data.PrisonerBook
 import uk.gov.justice.digital.hmpps.prisonerfromnomismigration.nomisprisoner.model.ProfileDetailsResponse
 
 @Service
-class ContactPersonBookingMovedService(
+class ContactPersonProfileDetailsBookingMovedService(
   private val nomisApiService: ContactPersonProfileDetailsNomisApiService,
-  private val nomisSyncApiService: ContactPersonNomisSyncApiService,
+  private val nomisSyncApiService: ContactPersonProfileDetailsNomisSyncApiService,
   private val syncService: ContactPersonProfileDetailsSyncService,
   private val telemetryClient: TelemetryClient,
 ) {
